requirements.txt
mysql~=0.0.3
mysql-connector-python~=9.3.0
python-dotenv~=1.1.0
uvicorn~=0.34.3
fastapi~=0.115.12
pydantic~=2.11.7
pytest~=8.4.0
boto3~=1.38.38
requests~=2.32.4
openai~=1.88.0
pillow~=11.2.1